+++ week/14/Week 14 - TSP/src/planarGraph/PlanarGraph.kt
@@ -1,7 +1,7 @@
 package planarGraph
 
 interface PlanarGraph {
-    fun distance(a: Int, b: Int): Double
+    fun distance(a: Int, b: Int): Float
     val size: Int
     val labels: IntArray
 }
